refactor: generalize file selection helper for consistency
- Move selectPdfFileFromWorkspace to dialogs module as selectFileFromWorkspace
- Make helper work with any file type, not just PDFs
- Update all image commands to use the generalized helper consistently
- Export FileSelectionResult interface for broader reuse
- Update tests to reflect new location and generalized approach
- Remove unused WorkspaceFS import from imageCommands

This improves code maintainability by following DRY principles and
ensuring consistent file selection patterns across all commands.
