+++ src/utils/dialogs.ts
@@ -69,3 +69,32 @@ export async function selectFile(
   const paths = await selectFiles({ ...options, allowMany: false });
   return paths ? paths[0] : null;
 }
+
+export interface FileSelectionResult {
+  relativePath: string;
+  absolutePath: string;
+}
+
+/**
+ * Prompts the user to select a file within the current workspace.
+ * Returns both the workspace-relative and absolute paths for downstream
+ * consumers. Returns null when the workspace is unavailable or the user
+ * cancels the dialog.
+ */
+export async function selectFileFromWorkspace(
+  options: FileDialogOptions,
+): Promise<FileSelectionResult | null> {
+  if (!WorkspaceFS.getPath()) {
+    vscode.window.showErrorMessage('No workspace folder open');
+    return null;
+  }
+
+  const relativePath = await selectFile(options);
+
+  if (!relativePath) {
+    return null;
+  }
+
+  const absolutePath = WorkspaceFS.fullPath(relativePath);
+  return { relativePath, absolutePath };
+}
